接口返回值修改

+++ kinbuffer/kinbuffer-core/src/main/java/org/kin/kinbuffer/io/ByteArrayInput.java
@@ -32,7 +32,7 @@ public final class ByteArrayInput implements Input{
     }
 
     @Override
-    public int readByte() {
+    public byte readByte() {
         return input.readByte();
     }
