+++ kinbuffer/kinbuffer-core/src/main/java/org/kin/kinbuffer/io/NioBufInput.java
@@ -43,7 +43,7 @@ import java.nio.charset.StandardCharsets;
     }
 
     @Override
-    public int readByte() {
+    public byte readByte() {
         return byteBuffer.get();
     }
